+++ codec2talkie/src/main/java/com/radio/codec2talkie/app/AppService.java
@@ -20,7 +20,6 @@ import android.os.Messenger;
 import android.os.PowerManager;
 import android.os.RemoteException;
 import android.util.Log;
-import android.util.Pair;
 import android.widget.Toast;
 
 import androidx.annotation.Nullable;
